# swagger-docs.json
{"openapi":"3.0.1","info":{"title":"Resa API","description":"API de réservation de traversées maritimes","contact":{"name":"Resa Team","email":"contact@resa.com"},"version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"User Management","description":"APIs for managing users"},{"name":"Boat Management","description":"APIs for managing boats"},{"name":"Travel Management","description":"APIs for managing travels"},{"name":"Authentication","description":"APIs for user authentication"}],"paths":{"/api/vehicles/{id}":{"get":{"tags":["vehicle-controller"],"operationId":"getVehicleById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Vehicle"}}}}}},"put":{"tags":["vehicle-controller"],"operationId":"updateVehicle","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Vehicle"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Vehicle"}}}}}},"delete":{"tags":["vehicle-controller"],"operationId":"deleteVehicle","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/users/{id}":{"get":{"tags":["User Management"],"summary":"Get user by ID","description":"Retrieves a user by their ID","operationId":"getUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}},"put":{"tags":["User Management"],"summary":"Update user","description":"Updates an existing user's information","operationId":"updateUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}},"delete":{"tags":["User Management"],"summary":"Delete user","description":"Deletes a user account","operationId":"deleteUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/users/{id}/role":{"put":{"tags":["User Management"],"summary":"Update user role","description":"Updates the role of an existing user","operationId":"updateUserRole","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"role","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/api/travels/{id}":{"get":{"tags":["Travel Management"],"summary":"Get travel by ID","description":"Retrieves a travel by its ID","operationId":"getTravelById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Travel"}}}}}},"put":{"tags":["Travel Management"],"summary":"Update travel","description":"Updates an existing travel's information","operationId":"updateTravel","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Travel"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Travel"}}}}}},"delete":{"tags":["Travel Management"],"summary":"Delete travel","description":"Deletes a travel","operationId":"deleteTravel","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/invoices/{id}":{"get":{"tags":["invoice-controller"],"operationId":"getInvoiceById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Invoice"}}}}}},"put":{"tags":["invoice-controller"],"operationId":"updateInvoice","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Invoice"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Invoice"}}}}}},"delete":{"tags":["invoice-controller"],"operationId":"deleteInvoice","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/extras/{id}":{"get":{"tags":["extra-controller"],"operationId":"getExtraById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Extra"}}}}}},"put":{"tags":["extra-controller"],"operationId":"updateExtra","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Extra"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Extra"}}}}}},"delete":{"tags":["extra-controller"],"operationId":"deleteExtra","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/boats/{id}":{"get":{"tags":["Boat Management"],"summary":"Get boat by ID","description":"Retrieves a boat by its ID","operationId":"getBoatById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Boat"}}}}}},"put":{"tags":["Boat Management"],"summary":"Update boat","description":"Updates an existing boat's information","operationId":"updateBoat","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Boat"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Boat"}}}}}},"delete":{"tags":["Boat Management"],"summary":"Delete boat","description":"Deletes a boat","operationId":"deleteBoat","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/vehicles":{"get":{"tags":["vehicle-controller"],"operationId":"getAllVehicles","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Vehicle"}}}}}}},"post":{"tags":["vehicle-controller"],"operationId":"createVehicle","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Vehicle"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Vehicle"}}}}}}},"/api/users":{"get":{"tags":["User Management"],"summary":"Get all users","description":"Retrieves a list of all users","operationId":"getAllUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}}},"post":{"tags":["User Management"],"summary":"Create new user","description":"Creates a new user account","operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/api/travels":{"get":{"tags":["Travel Management"],"summary":"Get all travels","description":"Retrieves a list of all travels","operationId":"getAllTravels","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Travel"}}}}}}},"post":{"tags":["Travel Management"],"summary":"Create travel","description":"Creates a new travel","operationId":"createTravel","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Travel"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Travel"}}}}}}},"/api/invoices":{"get":{"tags":["invoice-controller"],"operationId":"getAllInvoices","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Invoice"}}}}}}},"post":{"tags":["invoice-controller"],"operationId":"createInvoice","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Invoice"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Invoice"}}}}}}},"/api/invoices/{invoiceId}/extras":{"post":{"tags":["invoice-controller"],"operationId":"addExtraToInvoice","parameters":[{"name":"invoiceId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Extra"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Invoice"}}}}}}},"/api/extras":{"get":{"tags":["extra-controller"],"operationId":"getAllExtras","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Extra"}}}}}}},"post":{"tags":["extra-controller"],"operationId":"createExtra","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Extra"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Extra"}}}}}}},"/api/boats":{"get":{"tags":["Boat Management"],"summary":"Get all boats","description":"Retrieves a list of all boats","operationId":"getAllBoats","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Boat"}}}}}}},"post":{"tags":["Boat Management"],"summary":"Create boat","description":"Creates a new boat","operationId":"createBoat","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Boat"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Boat"}}}}}}},"/api/auth/register":{"post":{"tags":["Authentication"],"summary":"Register new user","description":"Creates a new user account and returns a JWT token","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/auth/login":{"post":{"tags":["Authentication"],"summary":"Login user","description":"Authenticates a user and returns a JWT token","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/vehicles/user/{userId}":{"get":{"tags":["vehicle-controller"],"operationId":"getVehiclesByUserId","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Vehicle"}}}}}}}},"/api/travels/departure-port/{port}":{"get":{"tags":["Travel Management"],"summary":"Get travels by departure port","description":"Retrieves a list of travels from the specified port","operationId":"getTravelsByDeparturePort","parameters":[{"name":"port","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Travel"}}}}}}}},"/api/travels/date-range":{"get":{"tags":["Travel Management"],"summary":"Get travels by date range","description":"Retrieves a list of travels within the specified date range","operationId":"getTravelsByDateRange","parameters":[{"name":"startDate","in":"query","required":true,"schema":{"type":"string","format":"date-time"}},{"name":"endDate","in":"query","required":true,"schema":{"type":"string","format":"date-time"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Travel"}}}}}}}},"/api/travels/boat/{boatId}":{"get":{"tags":["Travel Management"],"summary":"Get travels by boat","description":"Retrieves a list of travels for the specified boat","operationId":"getTravelsByBoat","parameters":[{"name":"boatId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Travel"}}}}}}}},"/api/travels/arrival-port/{port}":{"get":{"tags":["Travel Management"],"summary":"Get travels by arrival port","description":"Retrieves a list of travels to the specified port","operationId":"getTravelsByArrivalPort","parameters":[{"name":"port","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Travel"}}}}}}}},"/api/health":{"get":{"tags":["health-controller"],"operationId":"healthCheck","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/boats/departure-port/{port}":{"get":{"tags":["Boat Management"],"summary":"Get boats by departure port","description":"Retrieves a list of boats that have travels from the specified port","operationId":"getBoatsByDeparturePort","parameters":[{"name":"port","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Boat"}}}}}}}},"/api/boats/available":{"get":{"tags":["Boat Management"],"summary":"Get available boats","description":"Retrieves a list of boats that are not assigned to any travel","operationId":"getAvailableBoats","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Boat"}}}}}}}},"/api/boats/available/departure-port/{port}":{"get":{"tags":["Boat Management"],"summary":"Get available boats by departure port","description":"Retrieves a list of available boats that have travels from the specified port","operationId":"getAvailableBoatsByDeparturePort","parameters":[{"name":"port","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Boat"}}}}}}}},"/api/boats/available/arrival-port/{port}":{"get":{"tags":["Boat Management"],"summary":"Get available boats by arrival port","description":"Retrieves a list of available boats that have travels to the specified port","operationId":"getAvailableBoatsByArrivalPort","parameters":[{"name":"port","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Boat"}}}}}}}},"/api/boats/arrival-port/{port}":{"get":{"tags":["Boat Management"],"summary":"Get boats by arrival port","description":"Retrieves a list of boats that have travels to the specified port","operationId":"getBoatsByArrivalPort","parameters":[{"name":"port","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Boat"}}}}}}}},"/api/auth/me":{"get":{"tags":["Authentication"],"summary":"Get current user","description":"Retrieves information about the currently authenticated user","operationId":"getCurrentUser","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/invoices/{invoiceId}/extras/{extraId}":{"delete":{"tags":["invoice-controller"],"operationId":"removeExtraFromInvoice","parameters":[{"name":"invoiceId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"extraId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}}},"components":{"schemas":{"Boat":{"required":["cabins","name","placeFreight","placeHumanAndPet"],"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"placeHumanAndPet":{"type":"integer","format":"int32"},"cabins":{"type":"integer","format":"int32"},"placeFreight":{"type":"integer","format":"int32"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"travels":{"type":"array","items":{"$ref":"#/components/schemas/Travel"}}}},"Extra":{"required":["name","price","quantity"],"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"price":{"type":"number"},"quantity":{"type":"integer","format":"int32"},"travel":{"$ref":"#/components/schemas/Travel"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"GrantedAuthority":{"type":"object","properties":{"authority":{"type":"string"}}},"Invoice":{"required":["travel","user"],"type":"object","properties":{"id":{"type":"integer","format":"int32"},"travel":{"$ref":"#/components/schemas/Travel"},"user":{"$ref":"#/components/schemas/User"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"invoiceExtras":{"type":"array","items":{"$ref":"#/components/schemas/InvoiceExtra"}}}},"InvoiceExtra":{"required":["extra","invoice"],"type":"object","properties":{"id":{"type":"integer","format":"int32"},"invoice":{"$ref":"#/components/schemas/Invoice"},"extra":{"$ref":"#/components/schemas/Extra"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"Travel":{"required":["arrivalPort","arrivalTime","boat","departurePort","departureTime","priceHuman","priceVehicle"],"type":"object","properties":{"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int32"},"departureTime":{"type":"string","format":"date-time"},"arrivalTime":{"type":"string","format":"date-time"},"departurePort":{"maxLength":100,"minLength":2,"type":"string"},"arrivalPort":{"maxLength":100,"minLength":2,"type":"string"},"priceHuman":{"minimum":0.0,"exclusiveMinimum":true,"type":"number"},"priceVehicle":{"minimum":0.0,"exclusiveMinimum":true,"type":"number"},"boat":{"$ref":"#/components/schemas/Boat"},"extras":{"type":"array","items":{"$ref":"#/components/schemas/Extra"}},"invoices":{"type":"array","items":{"$ref":"#/components/schemas/Invoice"}},"users":{"type":"array","items":{"$ref":"#/components/schemas/User"}},"vehicles":{"type":"array","items":{"$ref":"#/components/schemas/Vehicle"}}}},"User":{"required":["email","name","password"],"type":"object","properties":{"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"roles":{"type":"array","items":{"$ref":"#/components/schemas/UserRoles"}},"vehicles":{"type":"array","items":{"$ref":"#/components/schemas/Vehicle"}},"userTravels":{"type":"array","items":{"$ref":"#/components/schemas/UserTravel"}},"invoices":{"type":"array","items":{"$ref":"#/components/schemas/Invoice"}},"travels":{"type":"array","items":{"$ref":"#/components/schemas/Travel"}},"enabled":{"type":"boolean"},"credentialsNonExpired":{"type":"boolean"},"accountNonExpired":{"type":"boolean"},"accountNonLocked":{"type":"boolean"},"authorities":{"type":"array","items":{"$ref":"#/components/schemas/GrantedAuthority"}},"username":{"type":"string"}}},"UserRoles":{"type":"object","properties":{"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int32"},"user":{"$ref":"#/components/schemas/User"},"roles":{"type":"string"}}},"UserTravel":{"type":"object","properties":{"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int32"},"user":{"$ref":"#/components/schemas/User"},"travel":{"$ref":"#/components/schemas/Travel"}}},"Vehicle":{"type":"object","properties":{"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int32"},"registration":{"type":"string"},"type":{"type":"string"},"user":{"$ref":"#/components/schemas/User"},"travels":{"type":"array","items":{"$ref":"#/components/schemas/Travel"}}}},"AuthDto":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}}}}}